requirements-dev.txt
coverage==4.0
coveralls==1.0
django-nose==1.4.2
flake8==2.3.0
flake8-import-order==0.5.3
flake8-pep257==1.0.3
https://github.com/zheller/flake8-quotes/tarball/aef86c4f8388e790332757e5921047ad53160a75#egg=flake8-quotes
nose==1.3.7
nosetests-json-extended==0.1.0
pep257==0.6.0
pep8==1.6.2
pep8-naming==0.3.3
pyflakes==1.0.0
